pyproject.toml

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "faststack"
version = "1.5.3"
authors = [
    { name="Alan Rockefeller"},
]
description = "Ultra-fast JPG Viewer for Focus Stacking Selection"
readme = "README.md"
requires-python = ">=3.11,<3.14"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: Microsoft :: Windows",
    "Operating System :: MacOS",
    "Operating System :: POSIX :: Linux",
]
dependencies = [
    "PySide6>=6.0,<7.0",
    "PyTurboJPEG>=1.8,<2.0",
    "numpy>=2.0,<3.0",
    "cachetools>=5.0,<6.0",
    "watchdog>=4.0,<5.0",
    "Pillow>=10.0,<11.0",
    "opencv-python>=4.0,<5.0",
]

[project.optional-dependencies]

dev = [
    "pytest>=8.0,<9.0",
    "build",
    "twine",
]

[project.scripts]
faststack = "faststack.app:cli"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["."]
include = ["faststack*"]

[tool.setuptools.package-data]
faststack = ["qml/**/*"]


[tool.pytest.ini_options]
testpaths = ["faststack/tests"]
python_files = ["test_*.py"]
addopts = "-p no:cacheprovider -p no:doctest --basetemp=./var/pytest-temp"
norecursedirs = ["var", ".venv", "cache", "faststack.egg-info", "__pycache__"]
pythonpath = ["."]

